Migrate Mutation.updateMemberRole
